requirements.txt
boto3 
python-dotenv 
qdrant-client 
openai 
langchain 
langchain-text-splitters 
pypdf 
fastapi 
uvicorn
PyMuPDF
fastapi==0.110.0
uvicorn==0.27.1
numpy==1.26.4
pydantic==2.6.4
qdrant-client==1.16.2
